requirements.txt
    pandas
numpy
yfinance
pyarrow
fastparquet
duckdb
PyYAML
streamlit
requests
